% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix Function}
\usage{
cor_matrix(dataset, n)
}
\arguments{
\item{n}{Input individual year. Example: 2000}
}
\description{
The input for this function should be an individual year.
The function should limit the data for that year to include numeric continuous variables only (not categorical data).
The output should be the correlation matrix for all numeric variables in the data for that year.
}
\examples{
cor_matrix_nba("dataset,2000")
}
