% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Oldest Player Function}
\usage{
oldest(dataset, n)
}
\arguments{
\item{n}{Input individual year. Example: 2000}
}
\description{
This function provides the oldest player in the NBA data for that year.
The input should be an individual year.
}
\examples{
oldest("dataset,2000")
}
