% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_10_player_pts.R
\name{top_10_player_pts}
\alias{top_10_player_pts}
\title{Year Top 10 NBA players Function}
\usage{
top_10_player_pts(dataset, n)
}
\arguments{
\item{n}{Input individual year. Example: 2000}
}
\value{
the name of the top 10 players
}
\description{
This function allows you to find the top 10 NBA players with highest points for a certain year.
Your input will be the year.
}
\examples{
top_10_player_pts("dataset,2000")
}
